backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
numpy==2.1.3
scipy==1.14.1
pyvista==0.44.2
google-generativeai==0.8.3
snowflake-connector-python==3.12.3
aiofiles==24.1.0
pydicom==3.0.1
scikit-image==0.24.0
Pillow==11.0.0
nibabel>=5.0.0
trimesh>=3.20.0
